parsing_testing/requirements.txt
argcomplete==1.8.2
beautifulsoup4==4.5.3
bkcharts==0.2
bokeh==0.12.6
boto==2.47.0
bz2file==0.98
certifi==2017.4.17
chardet==2.3.0
cycler==0.10.0
cymem==1.31.2
cytoolz==0.8.2
decorator==4.0.11
dill==0.2.6
docx2txt==0.6
EbookLib==0.15
en-core-web-sm==1.2.0
ftfy==4.4.3
gensim==2.2.0
html5lib==0.999999999
idna==2.5
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
Jinja2==2.9.6
jupyter-client==5.1.0
jupyter-core==4.3.0
lxml==3.8.0
MarkupSafe==1.0
matplotlib==2.0.2
murmurhash==0.26.4
numpy==1.13.0
olefile==0.44
pandas==0.20.2
pathlib==1.0.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.2.0
plac==0.9.6
pocketsphinx==0.1.3
preshed==1.0.0
prompt-toolkit==1.0.14
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
pytesseract==0.1.7
python-dateutil==2.6.0
python-docx==0.8.6
python-pptx==0.6.5
pytz==2017.2
PyYAML==3.12
pyzmq==16.0.2
regex==2017.4.5
requests==2.18.1
scikit-learn==0.18.2
scipy==0.19.1
simplegeneric==0.8.1
six==1.10.0
sklearn==0.0
smart-open==1.5.3
spacy==1.8.2
SpeechRecognition==3.6.3
termcolor==1.1.0
textract==1.6.1
thinc==6.5.2
toolz==0.8.2
tornado==4.5.1
tqdm==4.14.0
traitlets==4.3.2
ujson==1.35
urllib3==1.21.1
Wand==0.4.4
wcwidth==0.1.7
webencodings==0.5.1
wrapt==1.10.10
xlrd==1.0.0
XlsxWriter==0.9.8
